chore: add Core pkg
